+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/LoadRuleProviders.kt
@@ -14,40 +14,51 @@ private val LOGGER = KotlinLogging.logger {}.initKtLintKLogger()
 /**
  * Loads given list of paths to jar files. For files containing a [RuleSetProviderV2] class, get all [RuleProvider]s.
  */
-internal fun List<URL>.loadRuleProviders(debug: Boolean): Set<RuleProvider> =
-    this
+internal fun List<URL>.loadRuleProvidersByRuleSetId(debug: Boolean): Map<String, Set<RuleProvider>> =
+    getKtlintRulesets()
         .plus(
-            // Ensure that always at least one element exists in this list so that the rule sets provided by the KtLint
-            // CLI module itself will be found even in case no JAR files are specified
-            null,
+            getRuleProvidersFromCustomRuleSetJars(debug)
         )
-        // Remove JAR files which were provided multiple times
-        .distinct()
-        .flatMap { getRuleProvidersFromJar(it, debug).values }
-        .flatten()
-        .toSet()
 
-private fun getRuleProvidersFromJar(
-    url: URL?,
-    debug: Boolean,
-): Map<String, Set<RuleProvider>> {
-    if (url != null && debug) {
-        LOGGER.debug { "JAR ruleset provided with path \"${url.path}\"" }
-    }
-    return try {
+private fun getKtlintRulesets(): Map<String, Set<RuleProvider>> {
+    return loadRulesetsFrom()
+}
+
+
+private fun loadRulesetsFrom(url: URL? = null): Map<String, Set<RuleProvider>> =
+    try {
         ServiceLoader
             .load(
                 RuleSetProviderV2::class.java,
                 URLClassLoader(listOfNotNull(url).toTypedArray()),
-            ).filter {
-                // The KtLint-root (CLI) module includes the standard and experimental rule sets. When those rule sets
-                // are also included in the specified JAR (url != null) then ignore them.
-                url == null || it.id !in KTLINT_RULE_SETS
-            }.associate { ruleSetProviderV2 -> ruleSetProviderV2.id to ruleSetProviderV2.getRuleProviders() }
-            .also { ruleSetIdMap ->
-                if (url != null && ruleSetIdMap.isEmpty()) {
-                    LOGGER.warn {
-                        """
+            ).associate { ruleSetProviderV2 -> ruleSetProviderV2.id to ruleSetProviderV2.getRuleProviders() }
+    } catch (e: ServiceConfigurationError) {
+        LOGGER.warn { "Error while loading rule set JAR '$url':\n${e.printStackTrace()}" }
+        emptyMap()
+    }
+
+private fun List<URL>.getRuleProvidersFromCustomRuleSetJars(debug: Boolean): Map<String, Set<RuleProvider>> =
+    this
+        // Remove JAR files which were provided multiple times
+        .distinct()
+        .flatMap { getRuleProvidersFromCustomRuleSetJar(it, debug).entries }
+        .associate { it.key to it.value }
+
+private fun getRuleProvidersFromCustomRuleSetJar(
+    url: URL,
+    debug: Boolean,
+): Map<String, Set<RuleProvider>> {
+    if (debug) {
+        LOGGER.debug { "JAR ruleset provided with path \"${url.path}\"" }
+    }
+    return loadRulesetsFrom(url)
+        .filterKeys {
+            // Ignore the Ktlint rule sets when they are included in the custom rule set.
+            it !in KTLINT_RULE_SETS
+        }.also { ruleSetIdMap ->
+            if (ruleSetIdMap.isEmpty()) {
+                LOGGER.warn {
+                    """
                         JAR ${url.path}, provided as command line argument, does not contain a custom ruleset provider.
                             Check following:
                               - Does the jar contain an implementation of the RuleSetProviderV2 interface?
@@ -55,12 +66,9 @@ private fun getRuleProvidersFromJar(
                               - Is the resource file located in directory "src/main/resources/META-INF/services"?
                               - Does the resource file contain the fully qualified class name of the class implementing the RuleSetProviderV2 interface?
                         """.trimIndent()
-                    }
                 }
             }
-    } catch (e: ServiceConfigurationError) {
-        emptyMap()
-    }
+        }
 }
 
 private val KTLINT_RULE_SETS = listOf("standard", "experimental")